requirements/base.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
redis==5.0.4
celery==5.3.6
bcrypt==4.0.1
python-slugify==8.0.4
slowapi==0.1.9
authorizenet==1.1.6
python-multipart==0.0.9